Finished lab 3, just forgot to upload
Did a basic form. Did some input validation. Had to learn how to check PHP environment because one of the validators required php-intl extension on so that was new. Did a basic echo to output a message back to the view but will update at a later time so it looks prettier.